Refactoring: Optimización de código y mejoras en la estructura 🔧 Gio_Funcionalidad_topico
