+++ src/main/java/gio/apiforoalura/config/JwtUtils.java
@@ -13,10 +13,11 @@ import java.security.Key;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.concurrent.TimeUnit;
 import java.util.function.Function;
 
 @Service
-public class JwtService {
+public class JwtUtils {
 
     private static final String SECRET_KEY;
 
@@ -42,7 +43,7 @@ public class JwtService {
                 .setClaims(extraClaims)
                 .setSubject(userDetails.getUsername())
                 .setIssuedAt(new Date(System.currentTimeMillis()))
-                .setExpiration(new Date(System.currentTimeMillis()  + (1000 * 60 * 24)))
+                .setExpiration(new Date(System.currentTimeMillis()  + TimeUnit.HOURS.toMillis(200)))
                 .signWith(getSignKey(), SignatureAlgorithm.HS256)
                 .compact();
     }
