+++ src/main/java/gio/apiforoalura/services/impl/UserDetailsServiceImpl.java
@@ -1,0 +1,20 @@
+package gio.apiforoalura.services.impl;
+
+import gio.apiforoalura.repositories.UserRepository;
+import jakarta.persistence.EntityNotFoundException;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class UserDetailsServiceImpl implements UserDetailsService {
+    private final UserRepository userRepository;
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        return userRepository.findByUserName(username)
+                .orElseThrow(() -> new EntityNotFoundException("No se encontro ningun usuario con el username dado"));
+    }
+}
